Const construct schema compile time
Dart has the ability to construct complex const objects compile time,
which avoids any runtime overhead with regards to construction, and are
never considered for garbage collection.

This is a very powerful feature, and it is used in Flutter to construct
static widgets.

This PR explores the possibility of doing the same for our schema.

It introduces fly-weight property classes that are constructed compile
time per realm model, and combined in a const constructed schema object.

These fly-weights are specialized by property type (ie. value, object, or
list - set and map will come later), and help us deconstruct the static
type of the collection, to get the static type of its elements.

This is useful to push type safety much deeper into the realm SDK, and
allows us to avoid a fair amount runtime branching when reading from a
realm. (The current PR doesn't do the same for writing yet)

It also refactor how we handle dynamic properties, which are now all
accessed by get<T> including for lists, ie. get<List<T>>. The error
messages are also improved.

The original idea came about, when implementing support for collections
of nullable primitives, but was spun out into its own PR, as it is a
significant change on its own.